+++ renderer/units/JavaScript Runner/samples/IDS/List IDS Indivisible Characters.js
@@ -2,6 +2,19 @@
 const { characters } = require ('./lib/unicode/parsed-ids-data.js');
 const { characterToCodePoint } = require ('./lib/unicode/unicode.js');
 const { coreSet, core2020Set } = require ('./lib/unicode/parsed-unihan-data.js');
+function getSet (codePoint)
+{
+    let set = "Full";
+    if (coreSet.includes (codePoint))
+    {
+        set = "IICore";
+    }
+    else if (core2020Set.includes (codePoint))
+    {
+        set = "U-Core";
+    }
+    return set;
+}
 let indivisibleCharacters = [ ];
 for (let character in characters)
 {
@@ -11,16 +24,7 @@ for (let character in characters)
     {
         if (sequence.ids === character)
         {
-            let set = "Full";
-            if (coreSet.includes (codePoint))
-            {
-                set = "IICore";
-            }
-            else if (core2020Set.includes (codePoint))
-            {
-                set = "U-Core";
-            }
-            indivisibleCharacters.push (`${codePoint}\t${character}\t${set}`);
+            indivisibleCharacters.push (`${codePoint}\t${character}\t${getSet (codePoint)}`);
             break;
         }
     }